Fix for skill order influencing results.
The script originally failed to check if something could be concatenated in reverse order, for example with skill list:
skills = {
    "Demo Slayer Arhat": "TTSSOO",
    "Heart-burning Spell": "SST",
}
It failed to discover the pattern "SSTTSSOO".
Switching the two skills would have the original script work:
skills = {
    "Heart-burning Spell": "SST",
    "Demo Slayer Arhat": "TTSSOO",
}
This commit fixes this issue by checking forward and back ends of concatenation pairs.

Additionally, checking front and back allows for a more optimal solution.

This does make me ponder a better solution for pattern discovery, the script may be do for a rewrite at some point.